refactor(mir): canonicalize concrete types during lowering
- add a MIR-owned concrete type interner and scoped substitution environments
- monomorphize functions using concrete type IDs
- correctly resolve generic trait members and associated type projections
- split MIR lowering into focused builder, type, layout, mono, signature,
  intrinsic, place, and static-data modules
- rename MIR Type to ValueType and update downstream consumers
- add regression coverage for generic substitution and aggregate lowering

+++ crates/wx-compiler/src/tir/builder/generics.rs
@@ -516,7 +516,7 @@ impl<'ast> Builder<'ast, '_> {
 		if ast_params.is_empty() {
 			return;
 		}
-		let offset = self.inherited_type_param_count(owner);
+		let offset = self.inherited_type_param_count(owner) as usize;
 		for (i, tp) in ast_params.iter().enumerate() {
 			let resolved = tp
 				.bounds
@@ -575,7 +575,7 @@ impl<'ast> Builder<'ast, '_> {
 	pub(super) fn inherited_type_param_count(
 		&self,
 		owner: TypeParamOwner,
-	) -> usize {
+	) -> u32 {
 		match owner {
 			TypeParamOwner::Function(id) => {
 				self.items.function_index(id).map_or(0, |idx| {
